+++ src/main/java/com/part3/team07/sb01deokhugamteam07/repository/CommentRepository.java
@@ -5,6 +5,8 @@ import com.part3.team07.sb01deokhugamteam07.entity.Review;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
+
+import com.part3.team07.sb01deokhugamteam07.repository.querydsl.CommentRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, UUID>, CommentRepositoryCustom {
