+++ src/main/java/com/part3/team07/sb01deokhugamteam07/repository/LikeRepository.java
@@ -6,6 +6,8 @@ import java.util.List;
 import java.util.Optional;
 
 import java.util.UUID;
+
+import com.part3.team07.sb01deokhugamteam07.repository.querydsl.LikeRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, UUID>, LikeRepositoryCustom {
