+++ src/main/java/com/part3/team07/sb01deokhugamteam07/repository/ReviewRepository.java
@@ -8,6 +8,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 import com.part3.team07.sb01deokhugamteam07.entity.User;
+import com.part3.team07.sb01deokhugamteam07.repository.querydsl.ReviewRepositoryCustom;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
