+++ src/main/java/com/part3/team07/sb01deokhugamteam07/repository/querydsl/BookRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.part3.team07.sb01deokhugamteam07.repository;
+package com.part3.team07.sb01deokhugamteam07.repository.querydsl;
 
 import com.part3.team07.sb01deokhugamteam07.entity.Book;
 import java.time.LocalDateTime;
