+++ src/main/java/com/part3/team07/sb01deokhugamteam07/repository/querydsl/LikeRepositoryImpl.java
@@ -1,6 +1,5 @@
-package com.part3.team07.sb01deokhugamteam07.repository;
+package com.part3.team07.sb01deokhugamteam07.repository.querydsl;
 
-import com.part3.team07.sb01deokhugamteam07.entity.QLike;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
