+++ src/main/java/com/part3/team07/sb01deokhugamteam07/repository/querydsl/NotificationRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package com.part3.team07.sb01deokhugamteam07.repository;
+package com.part3.team07.sb01deokhugamteam07.repository.querydsl;
 
 import com.part3.team07.sb01deokhugamteam07.entity.Notification;
 import com.part3.team07.sb01deokhugamteam07.entity.QNotification;
