+++ src/main/java/com/part3/team07/sb01deokhugamteam07/service/CommentService.java
@@ -64,7 +64,7 @@ public class CommentService {
 
     commentRepository.save(comment);
     log.info("create comment complete: id={}, comment={}", comment.getId(), comment.getContent());
-    increaseCommentCount(review.getId()); //review commentCount 증가
+    increaseCommentCount(review); //review commentCount 증가
     return commentMapper.toDto(comment);
   }
 
@@ -225,20 +225,22 @@ public class CommentService {
   //commentCount 감소 - 논리삭제
   private void decreaseCommentCountOnSoftDelete(Comment comment) {
     log.debug("댓글 논리 삭제로 commentCount 감소: reviewId={}, commentId={}", comment.getReview().getId(), comment.getId());
-    reviewRepository.decrementCommentCount(comment.getReview().getId());
+    comment.getReview().decreaseCommentCount();
+    //reviewRepository.decrementCommentCount(comment.getReview().getId());
   }
 
   //commentCount 감소 - 물리삭제
   private void decreaseCommentCountOnHardDelete(Comment comment) {
     if (!comment.isDeleted()) {
       log.debug("댓글 물리 삭제로 commentCount 감소: reviewId={}, commentId={}", comment.getReview().getId(), comment.getId());
-      reviewRepository.decrementCommentCount(comment.getReview().getId());
+      comment.getReview().decreaseCommentCount();
+      //reviewRepository.decrementCommentCount(comment.getReview().getId());
     }
   }
 
   //commentCount 증가
-  private void increaseCommentCount(UUID reviewId) {
-    log.debug("댓글 생성으로 리뷰 commentCount 증가: reviewId={}", reviewId);
-    reviewRepository.incrementCommentCount(reviewId);
+  private void increaseCommentCount(Review review) {
+    log.debug("댓글 생성으로 리뷰 commentCount 증가: reviewId={}", review.getId());
+    review.increaseCommentCount();
   }
 }
