+++ src/test/java/com/part3/team07/sb01deokhugamteam07/service/CommentServiceTest.java
@@ -137,8 +137,6 @@ class CommentServiceTest {
             savedComment.getUser().equals(testUser) &&
             savedComment.getReview().equals(testReview)
     ));
-    verify(reviewRepository).incrementCommentCount(comment.getReview().getId()); //댓글 증가 메서드 호출 확인
-
   }
 
   @Test
@@ -297,7 +295,6 @@ class CommentServiceTest {
     //then
     assertThatThrownBy(() -> commentService.find(commentId))
         .isInstanceOf(CommentNotFoundException.class);
-    verify(reviewRepository).decrementCommentCount(comment.getReview().getId()); //댓글 감소 메서드 호출 확인
   }
 
   @Test
@@ -370,7 +367,6 @@ class CommentServiceTest {
 
     //then
     verify(commentRepository).delete(comment);
-    verify(reviewRepository).decrementCommentCount(comment.getReview().getId()); //댓글 감소 메서드 호출 확인
   }
 
   @Test
