+++ src/test/java/com/part3/team07/sb01deokhugamteam07/service/DashboardServiceTest.java
@@ -21,7 +21,7 @@ import com.part3.team07.sb01deokhugamteam07.entity.User;
 import com.part3.team07.sb01deokhugamteam07.entity.ValueType;
 import com.part3.team07.sb01deokhugamteam07.repository.BookRepository;
 import com.part3.team07.sb01deokhugamteam07.repository.DashboardRepository;
-import com.part3.team07.sb01deokhugamteam07.repository.DashboardRepositoryCustom;
+import com.part3.team07.sb01deokhugamteam07.repository.querydsl.DashboardRepositoryCustom;
 import com.part3.team07.sb01deokhugamteam07.repository.ReviewRepository;
 import com.part3.team07.sb01deokhugamteam07.repository.UserRepository;
 import java.math.BigDecimal;
